Refactor to use form controller for confirm and delete pages
* removed confirm and delete routers, form has been extended to handle these pages
