+++ pages/common/routers/form.js
@@ -1,6 +1,6 @@
 const { Router } = require('express');
 const bodyParser = require('body-parser');
-const { mapValues, size, some, get, isEqual } = require('lodash');
+const { mapValues, size, some, get, isEqual, merge } = require('lodash');
 const validator = require('../../../lib/validation');
 
 const defaultMiddleware = (req, res, next) => next();
@@ -21,6 +21,7 @@ const flattedNested = (data, schema) =>
 module.exports = ({
   schema = {},
   model = 'model',
+  cancelPath = '/',
   configure = defaultMiddleware,
   clearErrors = defaultMiddleware,
   getValues = defaultMiddleware,
@@ -28,7 +29,13 @@ module.exports = ({
   validate = defaultMiddleware,
   getValidationErrors = defaultMiddleware,
   locals = defaultMiddleware,
-  saveValues = defaultMiddleware
+  saveValues = defaultMiddleware,
+  cancelEdit = (req, res, next) => {
+    return res.redirect(cancelPath);
+  },
+  editAnswers = (req, res, next) => {
+    return res.redirect(req.baseUrl.replace(/\/confirm/, ''));
+  }
 } = {}) => {
   const form = Router();
 
@@ -40,6 +47,18 @@ module.exports = ({
     req.session.form = req.session.form || {};
     req.session.form[req.model.id] = req.session.form[req.model.id] || {};
     return configure(req, res, next);
+  };
+
+  const _processQuery = (req, res, next) => {
+    const { clear, edit } = req.query;
+    if (clear) {
+      delete req.session.form[req.model.id];
+      return cancelEdit(req, res, next);
+    }
+    if (edit) {
+      return editAnswers(req, res, next);
+    }
+    next();
   };
 
   const _clearErrors = (req, res, next) => {
@@ -63,7 +82,7 @@ module.exports = ({
   };
 
   const _checkChanged = (req, res, next) => {
-    if (some(req.form.values, (value, key) => hasChanged(value, key, req.model))) {
+    if (!size(req.form.schema) || some(req.form.values, (value, key) => hasChanged(value, key, req.model))) {
       return next();
     }
     return next({ validation: { form: 'unchanged' } });
@@ -78,7 +97,7 @@ module.exports = ({
   };
 
   const _saveValues = (req, res, next) => {
-    req.session.form[req.model.id].values = req.form.values;
+    merge(req.session.form[req.model.id], { values: req.form.values });
     return saveValues(req, res, next);
   };
 
@@ -88,8 +107,8 @@ module.exports = ({
   };
 
   const _locals = (req, res, next) => {
-    const { values, validationErrors, schema } = req.form;
-    Object.assign(res.locals.static, { errors: validationErrors, schema });
+    const { values, validationErrors } = req.form;
+    Object.assign(res.locals.static, { errors: validationErrors });
     Object.assign(res.locals, { model: values });
     return locals(req, res, next);
   };
@@ -97,7 +116,7 @@ module.exports = ({
   const errorHandler = (err, req, res, next) => {
     if (err.validation) {
       req.session.form[req.model.id].validationErrors = err.validation;
-      req.session.form[req.model.id].values = req.form.values;
+      merge(req.session.form[req.model.id], { values: req.form.values });
       return res.redirect(req.originalUrl);
     }
     return next(err);
@@ -105,6 +124,7 @@ module.exports = ({
 
   form.get('/',
     _configure,
+    _processQuery,
     _getValues,
     _getValidationErrors,
     _locals
