+++ pages/common/views/containers/model-summary.jsx
@@ -1,8 +1,9 @@
+import { pick } from 'lodash';
 import { connect } from 'react-redux';
 import ModelSummary from '../components/model-summary';
 
-const mapStateToProps = ({ model }, { formatters }) => ({
-  model,
+const mapStateToProps = ({ model, static: { schema } }, { formatters }) => ({
+  model: pick(model, Object.keys(schema)),
   formatters: formatters || {}
 });
