+++ pages/place/create/index.js
@@ -1,5 +1,7 @@
 const page = require('../../../lib/page');
-const add = require('../routers/add-or-edit');
+const amend = require('../routers/amend');
+const confirm = require('../routers/confirm');
+const success = require('../../common/routers/success');
 
 module.exports = settings => {
   const app = page({
@@ -8,12 +10,24 @@ module.exports = settings => {
     ...settings
   });
 
-  app.use(add());
+  app.use('/', amend());
 
-  app.use((req, res, next) => {
-    Object.assign(res.locals, { model: res.locals.static.values });
+  app.post('/', (req, res, next) => {
+    return res.redirect(`${req.baseUrl}/confirm`);
+  });
+
+  app.use('/confirm', confirm());
+
+  app.use('/confirm', (req, res, next) => {
+    res.locals.model = req.form.values;
     return next();
   });
+
+  app.post('/confirm', (req, res, next) => {
+    return res.redirect(req.originalUrl.replace(/\/confirm/, '/success'));
+  });
+
+  app.use('/success', success({ model: 'place' }));
 
   return app;
 };
