+++ pages/place/routers/confirm.js
@@ -1,0 +1,32 @@
+const form = require('../../common/routers/form');
+const { getEstablishment, getNacwoById } = require('../../common/helpers');
+const { schema } = require('../schema');
+
+module.exports = settings => form(Object.assign({
+  model: 'place',
+  schema: {
+    declaration: { validate: ['required'] }
+  },
+  locals: (req, res, next) => {
+    Object.assign(res.locals, { model: req.model });
+    Promise.all([
+      getEstablishment(req),
+      getNacwoById(req, req.form.values.nacwo)
+    ])
+      .then(([establishment, nacwo]) => {
+        Object.assign(res.locals.static, {
+          establishment,
+          schema,
+          values: {
+            ...req.session.form[req.model.id].values,
+            nacwo
+          }
+        });
+      })
+      .then(() => next())
+      .catch(next);
+  },
+  cancelEdit: (req, res, next) => {
+    return res.redirect(req.listPath);
+  }
+}, settings));
